Cache Lark parsers based on grammars, instead of SQLAlchemyBuilders based on tables
This makes the test suite run 0.5-1.5 seconds faster, and allows for sharing Lark parsers when two different tables have identical schemas (which is common in the test suite)

this isn't really that significant because we still aren't using a shared cache, but that is basically impossible with Lark + Earley grammars. We would have to switch to a LALR grammar (or just stop generating dynamic grammars) to really fix this.