[JB-5152] cache parsed ingredient fields and the result of their validation in the context of a specific grammar (#189)
* Cache Lark parsers based on grammars, instead of SQLAlchemyBuilders based on tables

This makes the test suite run 0.5-1.5 seconds faster, and allows for sharing Lark parsers when two different tables have identical schemas (which is common in the test suite)

this isn't really that significant because we still aren't using a shared cache, but that is basically impossible with Lark + Earley grammars. We would have to switch to a LALR grammar (or just stop generating dynamic grammars) to really fix this.

* add support for an explicit SQLAlchemy expression cache

in Juicebox we can pass in our redis-backed django cache and get the benefits of out-of-process caching for our SQLAlchemy expressions.

* load and save a single cache item for *all* expressions for a given table

while cache checks are pretty fast, they will randomly take 60-70ms. when you are doing one check for every ingredient in a table on every data service request, you are more likely to hit those slow cases. so we want to minimize the number of roundtrips to redis.

* sigh, I have been doing too much javascript

* add some comments about cache keys

* let's try including the table name in the expression cache

since I was getting errors about tables not existing in metadata during expression deserialization. So, we can't share SQLAlchemy expressions across tables even if the schemas & ingredients are identical, because those expressions have references to specific tables.

* trying one more thing: instead of caching the SQLAlchemy expression, just cache the tree and validator.

My suspicion is that the parsing is the most expensive thing, not the building of the SQLAlchemy expression.

This change allows us to share the cache even across different (identical) tables, since the tree doesn't encode anything specific to the table.

* add a couple of kinda dumb unit tests, and a benchmark-ish unit test

* be super defensive about dealing with cached data

- handle any exceptions from retrieving objects from the top-level cache, because these can cause e.g. pickle loading errors after upgrading versions of python/django/recipe
- also handle any exceptions while processing the result of a cache hit, because the data in there may be from an older version of recipe, and the data has many implementation details, not only of Recipe but also of Lark. If there are any errors, we just retry without the cache.

And I added a unit test that ensures that this error handling works.

* auto-format Python code

* add structlog to dependencies

* just define an extra_args tuple instead of a tree_to_expression closure

Co-authored-by: chris's formatting bot <chrisgemignani@users.noreply.github.com>