Ensure datatypes are always str
sometimes datatypes would be an instance of `Token`. This usually didn't cause a problem because *somehow* Token objects are JSON serializable (does lark monkeypatch the json module? I still haven't figured that out).

However, this is still a bug since we're leaking implementation details out of Recipe about its use of Lark, and it can cause problems if the user is trying to do something with a data type other than serialize it.